+++ gazon-service/src/main/resources/db/changelog/v-1.12/01-create-store-table.sql
@@ -1,0 +1,7 @@
+CREATE TABLE store (
+   id                BIGSERIAL PRIMARY KEY,
+   products          VARCHAR(255),
+   managers          VARCHAR(255),
+   owner_id          BIGINT NOT NULL,
+   FOREIGN KEY (owner_id) REFERENCES users(id)
+);