+++ gazon-service/src/main/resources/db/changelog/v-1.12/02-create-store_managers-table.sql
@@ -1,0 +1,7 @@
+CREATE TABLE store_managers (
+        store_id BIGINT,
+        managers_id BIGINT,
+
+        FOREIGN KEY (store_id) REFERENCES store(id) ON DELETE CASCADE,
+        FOREIGN KEY (managers_id) REFERENCES users(id) ON DELETE CASCADE
+);