+++ gazon-service/src/main/resources/db/changelog/v-1.12/04-fill-store_managers-table.sql
@@ -1,0 +1,4 @@
+INSERT INTO store_managers (store_id, managers_id)
+VALUES (1, 1);
+INSERT INTO store_managers (store_id, managers_id)
+VALUES (2, 2);